test: move SQL handlers out of conftest
