+++ tests/unit/conftest.py
@@ -12,7 +12,6 @@ import pkgutil
 import sys
 import warnings
 from contextlib import asynccontextmanager
-from datetime import datetime as dt
 from pathlib import Path
 from unittest.mock import AsyncMock, MagicMock
 
@@ -166,820 +165,6 @@ class MockState:
 
 
 # ---------------------------------------------------------------------------
-# Domain SQL handlers
-#
-# Signature: handler(sql_text: str, params: dict) -> MockResult | None
-# Return MockResult to handle the query, or None to pass to the next handler.
-#
-# Stateful handlers are factories: make_xxx_handler(state) -> handler
-# Stateless handlers are plain functions: xxx_handler(sql_text, params)
-# ---------------------------------------------------------------------------
-
-
-def make_customer_handler(state: MockState):
-    """Handle all customer-related SQL (INSERT, UPDATE, DELETE, SELECT, COUNT)."""
-
-    def handler(sql_text, params):
-        # INSERT
-        if "insert into customers" in sql_text:
-            cid = state.customers_next_id
-            state.customers_next_id += 1
-            tags_raw = params.get("tags", [])
-            tags_val = _json.dumps(tags_raw) if not isinstance(tags_raw, str) else tags_raw
-            record = {
-                "id": cid,
-                "tenant_id": params.get("tenant_id", 0),
-                "name": params.get("name", "Test"),
-                "email": params.get("email"),
-                "phone": params.get("phone"),
-                "company": params.get("company"),
-                "status": params.get("status", "lead"),
-                "owner_id": params.get("owner_id", 0),
-                "tags": tags_val,
-                "created_at": params.get("created_at"),
-                "updated_at": params.get("updated_at"),
-            }
-            state.customers[cid] = record
-            return MockResult([MockRow(record.copy())])
-
-        # UPDATE
-        if sql_text.startswith("update") and "customers" in sql_text:
-            customer_id = params.get("id")
-            if customer_id in state.customers:
-                rec = state.customers[customer_id]
-                for k, v in params.items():
-                    if k not in ("id", "tenant_id"):
-                        rec[k] = v
-                return MockResult([MockRow(rec.copy())])
-            if customer_id and customer_id >= 1:
-                return MockResult(
-                    [
-                        MockRow(
-                            {
-                                "id": customer_id,
-                                "tenant_id": params.get("tenant_id", 1),
-                                "name": params.get("name", "Customer A"),
-                                "email": params.get("email", "a@test.com"),
-                                "phone": params.get("phone", "123"),
-                                "company": params.get("company", "Acme"),
-                                "status": params.get("status", "lead"),
-                                "owner_id": params.get("owner_id", 1),
-                                "tags": _json.dumps(params.get("tags", [])),
-                                "created_at": None,
-                                "updated_at": None,
-                            }
-                        )
-                    ]
-                )
-            return MockResult([])
-
-        # DELETE
-        if sql_text.startswith("delete") and "customers" in sql_text:
-            return MockResult([MockRow({"id": params.get("id", 1)})])
-
-        # COUNT
-        if "select" in sql_text and "count" in sql_text and "from customers" in sql_text:
-            tenant_id = params.get("tenant_id")
-            count_val = 3 if tenant_id == 1 else 7
-            return MockResult([[count_val]])
-
-        # SELECT tags by id (more specific — check before generic SELECT by id)
-        if "select" in sql_text and "from customers" in sql_text and "tags" in sql_text and "where id" in sql_text:
-            customer_id = params.get("id")
-            if customer_id in state.customers:
-                c = state.customers[customer_id]
-                return MockResult([MockRow({"id": customer_id, "tags": c.get("tags", "[]")})])
-            return MockResult([])
-
-        # SELECT by id
-        if "from customers where id" in sql_text:
-            customer_id = params.get("id")
-            if customer_id in state.customers:
-                return MockResult([MockRow(state.customers[customer_id].copy())])
-            fixtures = {
-                1: {
-                    "id": 1,
-                    "tenant_id": 1,
-                    "name": "Customer A",
-                    "email": "a@test.com",
-                    "phone": "123",
-                    "company": "Acme",
-                    "status": "lead",
-                    "owner_id": 1,
-                    "tags": "[]",
-                    "created_at": None,
-                    "updated_at": None,
-                },
-            }
-            if customer_id in fixtures:
-                return MockResult([MockRow(fixtures[customer_id].copy())])
-            return MockResult([])
-
-        # SELECT list (no WHERE id)
-        if "select" in sql_text and "from customers" in sql_text and "where id" not in sql_text:
-            tenant_filter = params.get("tenant_id", 0)
-            rows = []
-            for cid, rec in state.customers.items():
-                if rec.get("tenant_id") == tenant_filter:
-                    rows.append(MockRow(rec.copy()))
-            rows.extend(
-                [
-                    MockRow(
-                        {
-                            "id": 1,
-                            "tenant_id": tenant_filter,
-                            "name": "Customer A",
-                            "email": "a@test.com",
-                            "phone": "123",
-                            "company": "Acme",
-                            "status": "lead",
-                            "owner_id": 1,
-                            "tags": "[]",
-                            "created_at": None,
-                            "updated_at": None,
-                        }
-                    ),
-                    MockRow(
-                        {
-                            "id": 2,
-                            "tenant_id": tenant_filter,
-                            "name": "Customer B",
-                            "email": "b@test.com",
-                            "phone": "456",
-                            "company": "Beta",
-                            "status": "customer",
-                            "owner_id": 1,
-                            "tags": "[]",
-                            "created_at": None,
-                            "updated_at": None,
-                        }
-                    ),
-                ]
-            )
-            if tenant_filter == 2:
-                rows = []
-            return MockResult(rows)
-
-        return None
-
-    return handler
-
-
-def make_user_handler(state: MockState):
-    """Handle all user-related SQL (INSERT, UPDATE, DELETE, SELECT, COUNT)."""
-
-    def handler(sql_text, params):
-        # INSERT (checked before the guard because INSERT doesn't have "from users")
-        if "insert into users" in sql_text:
-            uid = state.users_next_id
-            state.users_next_id += 1
-            record = {
-                "id": uid,
-                "tenant_id": params.get("tenant_id", 0),
-                "username": params.get("username"),
-                "email": params.get("email"),
-                "password_hash": None,
-                "role": params.get("role", "user"),
-                "status": "pending",
-                "full_name": params.get("full_name"),
-                "bio": None,
-                "created_at": params.get("created_at"),
-                "updated_at": params.get("created_at"),
-            }
-            state.users[uid] = record
-            return MockResult([MockRow(record.copy())])
-
-        # Guard: only handle queries that reference the users table
-        is_user_query = "from users" in sql_text or (sql_text.startswith("update") and "users" in sql_text)
-        if not is_user_query:
-            return None
-
-        # DELETE
-        if sql_text.startswith("delete") and "users" in sql_text:
-            user_id = params.get("id")
-            was_in_mock = user_id in state.users
-            if was_in_mock:
-                state.users[user_id]["status"] = "deleted"
-                del state.users[user_id]
-            state.deleted_user_ids.add(user_id)
-            if was_in_mock:
-                return MockResult([MockRow({"id": user_id})])
-            return MockResult([])
-
-        # UPDATE by id
-        if ("set " in sql_text or sql_text.startswith("update")) and "where id" in sql_text:
-            user_id = params.get("id")
-            if user_id in state.users:
-                for k, v in params.items():
-                    if k not in ("id", "tenant_id"):
-                        state.users[user_id][k] = v
-                return MockResult([MockRow(state.users[user_id].copy())])
-            return MockResult([])
-
-        # COUNT
-        if "select" in sql_text and "count" in sql_text and "from users" in sql_text:
-            count_val = len(state.users)
-            if count_val == 0:
-                tenant_id = params.get("tenant_id")
-                count_val = 7 if tenant_id == 1 else 3
-            return MockResult([[count_val]])
-
-        # SELECT by username
-        if "where username" in sql_text:
-            username = params.get("username")
-            for uid, rec in state.users.items():
-                if rec.get("username") == username:
-                    return MockResult([MockRow(rec.copy())])
-            fixtures = {
-                "existing": {
-                    "id": 1,
-                    "tenant_id": 0,
-                    "username": "existing",
-                    "email": "existing@test.com",
-                    "password_hash": None,
-                    "role": "user",
-                    "status": "pending",
-                    "full_name": None,
-                    "bio": None,
-                },
-                "alice": {
-                    "id": 1,
-                    "tenant_id": 1,
-                    "username": "alice",
-                    "email": "alice@test.com",
-                    "password_hash": None,
-                    "role": "user",
-                    "status": "active",
-                    "full_name": "Alice",
-                    "bio": "bio",
-                },
-                "john_doe": {
-                    "id": 2,
-                    "tenant_id": 0,
-                    "username": "john_doe",
-                    "email": "john@test.com",
-                    "password_hash": None,
-                    "role": "user",
-                    "status": "pending",
-                    "full_name": "John Doe",
-                    "bio": None,
-                },
-                "jane_doe": {
-                    "id": 3,
-                    "tenant_id": 0,
-                    "username": "jane_doe",
-                    "email": "jane@test.com",
-                    "password_hash": None,
-                    "role": "user",
-                    "status": "pending",
-                    "full_name": "Jane Doe",
-                    "bio": None,
-                },
-            }
-            if username in fixtures:
-                row = fixtures[username].copy()
-                now = dt.utcnow()
-                row["created_at"] = now
-                row["updated_at"] = now
-                return MockResult([MockRow(row)])
-            return MockResult([])
-
-        # SELECT by email
-        if "where email" in sql_text:
-            email = params.get("email")
-            for uid, rec in state.users.items():
-                if rec.get("email") == email:
-                    return MockResult([MockRow(rec.copy())])
-            if email == "existing@test.com":
-                now = dt.utcnow()
-                return MockResult(
-                    [
-                        MockRow(
-                            {
-                                "id": 1,
-                                "tenant_id": 0,
-                                "username": "existing",
-                                "email": "existing@test.com",
-                                "password_hash": None,
-                                "role": "user",
-                                "status": "pending",
-                                "full_name": None,
-                                "bio": None,
-                                "created_at": now,
-                                "updated_at": now,
-                            }
-                        )
-                    ]
-                )
-            return MockResult([])
-
-        # SELECT by id
-        if "where id" in sql_text:
-            user_id = params.get("id")
-            if user_id in state.users:
-                return MockResult([MockRow(state.users[user_id].copy())])
-            if user_id in state.deleted_user_ids:
-                return MockResult([])
-            fixtures = {
-                1: {
-                    "id": 1,
-                    "tenant_id": 0,
-                    "username": "existing",
-                    "email": "existing@test.com",
-                    "password_hash": None,
-                    "role": "user",
-                    "status": "pending",
-                    "full_name": None,
-                    "bio": None,
-                },
-                5: {
-                    "id": 5,
-                    "tenant_id": 1,
-                    "username": "charlie",
-                    "email": "charlie@test.com",
-                    "password_hash": None,
-                    "role": "admin",
-                    "status": "active",
-                    "full_name": "Charlie",
-                    "bio": "dev",
-                },
-            }
-            if user_id in fixtures:
-                row = fixtures[user_id].copy()
-                now = dt.utcnow()
-                row["created_at"] = now
-                row["updated_at"] = now
-                return MockResult([MockRow(row)])
-            return MockResult([])
-
-        # SELECT list (generic)
-        if state.users:
-            return MockResult([MockRow(r.copy()) for r in state.users.values()])
-        now = dt.utcnow()
-        rows = [
-            MockRow(
-                {
-                    "id": 1,
-                    "tenant_id": 0,
-                    "username": "existing",
-                    "email": "existing@test.com",
-                    "password_hash": None,
-                    "role": "user",
-                    "status": "pending",
-                    "full_name": None,
-                    "bio": None,
-                    "created_at": now,
-                    "updated_at": now,
-                }
-            ),
-            MockRow(
-                {
-                    "id": 2,
-                    "tenant_id": 0,
-                    "username": "john_doe",
-                    "email": "john@test.com",
-                    "password_hash": None,
-                    "role": "user",
-                    "status": "pending",
-                    "full_name": "John Doe",
-                    "bio": None,
-                    "created_at": now,
-                    "updated_at": now,
-                }
-            ),
-            MockRow(
-                {
-                    "id": 3,
-                    "tenant_id": 0,
-                    "username": "jane_doe",
-                    "email": "jane@test.com",
-                    "password_hash": None,
-                    "role": "user",
-                    "status": "pending",
-                    "full_name": "Jane Doe",
-                    "bio": None,
-                    "created_at": now,
-                    "updated_at": now,
-                }
-            ),
-        ]
-        return MockResult(rows)
-
-    return handler
-
-
-def tenant_handler(sql_text, params):
-    """Handle all tenant-related SQL (stateless)."""
-
-    # INSERT
-    if "insert into tenants" in sql_text:
-        row = MockRow(
-            {
-                "id": 1,
-                "name": params.get("name"),
-                "plan": params.get("plan"),
-                "status": "active",
-                "settings": params.get("settings", "{}"),
-                "created_at": params.get("now"),
-                "updated_at": params.get("now"),
-            }
-        )
-        return MockResult([row])
-
-    # DELETE
-    if "delete" in sql_text and "tenants" in sql_text:
-        tenant_id = params.get("tenant_id")
-        if tenant_id and tenant_id != 1:
-            return MockResult([])
-        return MockResult([[1, "Deleted Tenant", "pro", "deleted", "{}", None, None]])
-
-    # UPDATE
-    if "update" in sql_text and "tenants" in sql_text:
-        tenant_id = params.get("tenant_id") or params.get("id")
-        if tenant_id == 1:
-            return MockResult([[1, "Updated Name", "pro", "active", "{}", None, None]])
-        return MockResult([])
-
-    # COUNT
-    if "select" in sql_text and "count" in sql_text and "from tenants" in sql_text:
-        return MockResult([[2]])
-
-    # SELECT (not count)
-    if "select" in sql_text and "from tenants" in sql_text and "count" not in sql_text:
-        tenant_id = params.get("tenant_id")
-
-        # Specific lookup: LIMIT 1 + WHERE id
-        if "limit 1" in sql_text and "where id" in sql_text:
-            if tenant_id == 1:
-                return MockResult([MockRow({"id": 1})])
-            if tenant_id == 42:
-                return MockResult(
-                    [
-                        MockRow(
-                            {
-                                "id": 42,
-                                "name": "Beta Org",
-                                "plan": "enterprise",
-                                "status": "active",
-                                "settings": '{"sso": true}',
-                                "created_at": None,
-                                "updated_at": None,
-                            }
-                        )
-                    ]
-                )
-            return MockResult([])
-
-        # List query (no WHERE id)
-        if "where id" not in sql_text:
-            rows = [
-                MockRow(
-                    {
-                        "id": 1,
-                        "name": "Tenant A",
-                        "plan": "pro",
-                        "status": "active",
-                        "settings": "{}",
-                        "created_at": None,
-                        "updated_at": None,
-                    }
-                ),
-                MockRow(
-                    {
-                        "id": 2,
-                        "name": "Tenant B",
-                        "plan": "enterprise",
-                        "status": "active",
-                        "settings": "{}",
-                        "created_at": None,
-                        "updated_at": None,
-                    }
-                ),
-            ]
-            return MockResult(rows)
-
-        # General SELECT with WHERE id (not LIMIT 1)
-        if tenant_id == 42:
-            return MockResult(
-                [
-                    MockRow(
-                        {
-                            "id": 42,
-                            "name": "Beta Org",
-                            "plan": "enterprise",
-                            "status": "active",
-                            "settings": '{"sso": true}',
-                            "created_at": None,
-                            "updated_at": None,
-                        }
-                    )
-                ]
-            )
-        return MockResult([])
-
-    return None
-
-
-def pipeline_handler(sql_text, params):
-    """Handle pipeline-related SQL (stateless)."""
-
-    if "insert into pipelines" in sql_text:
-        return MockResult(
-            [
-                MockRow(
-                    {
-                        "id": 1,
-                        "tenant_id": params.get("tenant_id", 0),
-                        "name": params.get("name", "Pipeline"),
-                        "description": params.get("description"),
-                        "is_active": True,
-                        "created_at": params.get("created_at"),
-                        "updated_at": params.get("updated_at"),
-                    }
-                )
-            ]
-        )
-
-    if "from pipelines" in sql_text:
-        return MockResult(
-            [
-                MockRow(
-                    {
-                        "id": 1,
-                        "tenant_id": 1,
-                        "name": "Pipeline A",
-                        "description": "Desc",
-                        "is_active": True,
-                        "created_at": None,
-                        "updated_at": None,
-                    }
-                )
-            ]
-        )
-
-    return None
-
-
-def opportunity_handler(sql_text, params):
-    """Handle opportunity-related SQL (stateless)."""
-
-    if "insert into opportunities" in sql_text:
-        return MockResult(
-            [
-                MockRow(
-                    {
-                        "id": 1,
-                        "tenant_id": params.get("tenant_id", 0),
-                        "customer_id": 1,
-                        "title": params.get("title", "Opportunity"),
-                        "amount": params.get("amount", 1000),
-                        "stage": "qualification",
-                        "probability": 20,
-                        "owner_id": params.get("owner_id", 0),
-                        "expected_close_date": params.get("expected_close_date"),
-                        "created_at": params.get("created_at"),
-                        "updated_at": params.get("updated_at"),
-                    }
-                )
-            ]
-        )
-
-    if "from opportunities" in sql_text:
-        return MockResult(
-            [
-                MockRow(
-                    {
-                        "id": 1,
-                        "tenant_id": 1,
-                        "customer_id": 1,
-                        "title": "Opportunity A",
-                        "amount": 1000,
-                        "stage": "qualification",
-                        "probability": 20,
-                        "owner_id": 1,
-                        "expected_close_date": None,
-                        "created_at": None,
-                        "updated_at": None,
-                    }
-                )
-            ]
-        )
-
-    return None
-
-
-def ticket_sql_handler(sql_text, params):
-    """Handle ticket-related SQL (stateless)."""
-
-    if "insert into tickets" in sql_text:
-        return MockResult(
-            [
-                MockRow(
-                    {
-                        "id": 1,
-                        "tenant_id": params.get("tenant_id", 0),
-                        "subject": params.get("subject", "Ticket"),
-                        "description": params.get("description"),
-                        "status": "open",
-                        "priority": params.get("priority", "medium"),
-                        "customer_id": 1,
-                        "assignee_id": params.get("assignee_id"),
-                        "created_at": params.get("created_at"),
-                        "updated_at": params.get("updated_at"),
-                    }
-                )
-            ]
-        )
-
-    if "from tickets" in sql_text:
-        return MockResult(
-            [
-                MockRow(
-                    {
-                        "id": 1,
-                        "tenant_id": 1,
-                        "subject": "Issue A",
-                        "description": "Desc",
-                        "status": "open",
-                        "priority": "medium",
-                        "customer_id": 1,
-                        "assignee_id": 1,
-                        "created_at": None,
-                        "updated_at": None,
-                    }
-                )
-            ]
-        )
-
-    return None
-
-
-def campaign_handler(sql_text, params):
-    """Handle campaign-related SQL (stateless)."""
-
-    if "insert into campaigns" in sql_text:
-        return MockResult(
-            [
-                MockRow(
-                    {
-                        "id": 1,
-                        "tenant_id": params.get("tenant_id", 0),
-                        "name": params.get("name", "Campaign"),
-                        "campaign_type": params.get("campaign_type", "email"),
-                        "status": "draft",
-                        "created_by": params.get("created_by"),
-                        "created_at": params.get("created_at"),
-                        "updated_at": params.get("updated_at"),
-                    }
-                )
-            ]
-        )
-
-    if "from campaigns" in sql_text:
-        return MockResult(
-            [
-                MockRow(
-                    {
-                        "id": 1,
-                        "tenant_id": 1,
-                        "name": "Campaign A",
-                        "campaign_type": "email",
-                        "status": "draft",
-                        "created_by": 1,
-                        "created_at": None,
-                        "updated_at": None,
-                    }
-                )
-            ]
-        )
-
-    return None
-
-
-def make_activity_handler(state: MockState):
-    """Handle all activity-related SQL (INSERT, UPDATE, DELETE, SELECT, COUNT)."""
-
-    def param(params, name, default=None):
-        if name in params:
-            return params[name]
-        prefix = f"{name}_"
-        for key, value in params.items():
-            if key.startswith(prefix):
-                return value
-        return default
-
-    def matches_filters(record, tenant_id=None, customer_id=None, activity_type=None):
-        if tenant_id is not None and record.get("tenant_id") != tenant_id:
-            return False
-        if customer_id is not None and record.get("customer_id") != customer_id:
-            return False
-        if activity_type is not None and record.get("type") != activity_type:
-            return False
-        return True
-
-    def handler(sql_text, params):
-        # INSERT
-        if "insert into activities" in sql_text:
-            aid = state.activities_next_id
-            state.activities_next_id += 1
-            record = {
-                "id": aid,
-                "tenant_id": params.get("tenant_id", 0),
-                "customer_id": params.get("customer_id", 0),
-                "opportunity_id": params.get("opportunity_id"),
-                "type": params.get("type", "call"),
-                "content": params.get("content", ""),
-                "created_by": params.get("created_by", 0),
-                "created_at": params.get("created_at"),
-            }
-            state.activities[aid] = record
-            return MockResult([MockRow(record.copy())])
-
-        # Guard: only handle queries that reference the activities table
-        is_activity_query = (
-            "from activities" in sql_text
-            or (sql_text.startswith("update") and "activities" in sql_text)
-            or (sql_text.startswith("delete") and "activities" in sql_text)
-        )
-        if not is_activity_query:
-            return None
-
-        # UPDATE
-        if sql_text.startswith("update") and "activities" in sql_text:
-            activity_id = param(params, "id")
-            tenant_id = param(params, "tenant_id")
-            if activity_id in state.activities:
-                rec = state.activities[activity_id]
-                if tenant_id is not None and rec.get("tenant_id") != tenant_id:
-                    return MockResult([])
-                for k, v in params.items():
-                    if k not in ("id", "tenant_id") and not k.startswith(("id_", "tenant_id_")):
-                        rec[k] = v
-                return MockResult([MockRow(rec.copy())])
-            return MockResult([])
-
-        # DELETE
-        if sql_text.startswith("delete") and "activities" in sql_text:
-            activity_id = param(params, "id")
-            tenant_id = param(params, "tenant_id")
-            if activity_id in state.activities and matches_filters(state.activities[activity_id], tenant_id=tenant_id):
-                state.activities.pop(activity_id)
-                return MockResult([MockRow({"id": activity_id})])
-            return MockResult([])
-
-        # COUNT
-        if "select" in sql_text and "count" in sql_text and "from activities" in sql_text:
-            tenant_id = param(params, "tenant_id")
-            customer_id = param(params, "customer_id")
-            activity_type = param(params, "type")
-            count_val = len(
-                [
-                    r
-                    for r in state.activities.values()
-                    if matches_filters(r, tenant_id=tenant_id, customer_id=customer_id, activity_type=activity_type)
-                ]
-            )
-            return MockResult([[count_val]])
-
-        # SELECT by id
-        activity_id = param(params, "id")
-        if "where" in sql_text and "activities.id" in sql_text and activity_id is not None:
-            tenant_id = param(params, "tenant_id")
-            if activity_id in state.activities and matches_filters(state.activities[activity_id], tenant_id=tenant_id):
-                return MockResult([MockRow(state.activities[activity_id].copy())])
-            return MockResult([])
-
-        # SELECT list
-        if state.activities:
-            tenant_id = param(params, "tenant_id", 0)
-            customer_id = param(params, "customer_id")
-            activity_type = param(params, "type")
-            rows = [
-                MockRow(r.copy())
-                for r in state.activities.values()
-                if matches_filters(r, tenant_id=tenant_id, customer_id=customer_id, activity_type=activity_type)
-            ]
-            return MockResult(rows)
-
-        return MockResult([])
-
-    return handler
-
-
-def make_count_handler(state: MockState):
-    """Fallback COUNT handler for queries not caught by domain handlers."""
-
-    def handler(sql_text, params):
-        if "select" not in sql_text or "count" not in sql_text:
-            return None
-        tenant_id = params.get("tenant_id")
-        count_val = 7 if tenant_id == 1 else 3
-        return MockResult([[count_val]])
-
-    return handler
-
-
-# ---------------------------------------------------------------------------
 # Session builder
 # ---------------------------------------------------------------------------
 
@@ -989,8 +174,11 @@ def _load_domain_handler_modules():
     package_name = "tests.unit.domain_handlers"
     package = importlib.import_module(package_name)
     modules = []
-    for info in sorted(pkgutil.iter_modules(package.__path__, prefix=f"{package_name}."), key=lambda item: item.name):
+    discovered = []
+    for info in pkgutil.iter_modules(package.__path__, prefix=f"{package_name}."):
         module = importlib.import_module(info.name)
+        discovered.append(module)
+    for module in sorted(discovered, key=lambda item: (getattr(item, "ORDER", 100), item.__name__)):
         modules.append(module)
         for name in getattr(module, "__all__", []):
             if name != "get_handlers":
@@ -1012,18 +200,7 @@ def _domain_handlers(state: MockState):
 
 def all_handlers(state: MockState):
     """Return the full list of domain handlers (default for backward compat)."""
-    return [
-        make_customer_handler(state),
-        make_user_handler(state),
-        tenant_handler,
-        pipeline_handler,
-        opportunity_handler,
-        ticket_sql_handler,
-        campaign_handler,
-        make_activity_handler(state),
-        *_domain_handlers(state),
-        make_count_handler(state),
-    ]
+    return _domain_handlers(state)
 
 
 def make_mock_session(handlers=None):
